+++ lib/src/widgets/flutter_ddi_custom_pop_scope.dart
@@ -1,4 +1,3 @@
-import 'package:dart_ddi/dart_ddi.dart';
 import 'package:flutter/material.dart';
 
 /// Custom PopScope widget that handles module destruction when navigating back.
@@ -11,7 +10,6 @@ class CustomPopScope extends StatelessWidget {
   /// [onPopInvoked] - Callback when pop is invoked.
   const CustomPopScope({
     required this.child,
-    required this.moduleQualifier,
     required this.onPopInvoked,
     super.key,
   });
@@ -19,34 +17,19 @@ class CustomPopScope extends StatelessWidget {
   /// The child widget to wrap.
   final Widget child;
 
-  /// The qualifier of the module to destroy.
-  final Object moduleQualifier;
-
   /// Callback when pop is invoked.
-  final void Function(bool isDestroyed) onPopInvoked;
+  final Future<void> Function(bool isDestroyed) onPopInvoked;
 
   @override
   Widget build(BuildContext context) {
     return PopScope(
       canPop: false,
       onPopInvokedWithResult: (pop, result) async {
-        final navigator = Navigator.of(context);
-        if (!navigator.canPop() ||
-            pop && !ddi.isRegistered(qualifier: moduleQualifier)) {
-          return;
-        }
-
         try {
-          await ddi.destroy(qualifier: moduleQualifier);
-
-          onPopInvoked(true);
-
-          if (context.mounted && !pop) {
-            navigator.pop(result);
-          }
+          await onPopInvoked(true);
         } catch (e) {
           // Log error but don't throw to prevent app crashes
-          debugPrint('Error destroying module $moduleQualifier: $e');
+          debugPrint('Error destroying module: $e');
         }
       },
       child: child,